% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_network_props}
\alias{build_network_props}
\title{Construct visNetwork ready pokemon data}
\usage{
build_network_props(tmp, poke_data, i)
}
\arguments{
\item{poke_data}{Pokemon data list.}
}
\value{
A list.
}
\description{
Return a list of 2 dataframes, one for edges another for nodes.
}
\keyword{internal}
