% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_poke_families}
\alias{build_poke_families}
\title{Construct visNetwork pokemon families}
\usage{
build_poke_families(poke_data)
}
\arguments{
\item{poke_data}{Pokemon data list. Passed to \link{build_network_props}.}
}
\value{
A list.
}
\description{
Return a list of nodes and edges grouped by evolution.
}
\keyword{internal}
