% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeFightModule.R
\name{calculate_damages}
\alias{calculate_damages}
\title{Function that calculates the damages of an attack on a given opponent}
\usage{
calculate_damages(current_attack, current_pokemon, opponent, types)
}
\arguments{
\item{current_attack}{The currently selected attack.}

\item{current_pokemon}{Id of the randomly generated pokemon who is attacking.}

\item{opponent}{Opponent type. Useful for effectiveness calculations.}

\item{types}{Object containing all pokemons types strenght and weaknesses.}
}
\description{
Function that calculates the damages of an attack on a given opponent
}
\note{
As already explained, I only consider physical attacks...
}
