% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeFightModule.R
\name{compute_stat}
\alias{compute_stat}
\title{Function that calculate a given stat of a pokemon
based on its level.}
\usage{
compute_stat(level, base_stat)
}
\arguments{
\item{level}{Current pokemon level.}

\item{base_stat}{Stat to be scaled.}
}
\description{
Function that calculate a given stat of a pokemon
based on its level.
}
\note{
To be used in \link{generate_pokemons}.
}
