% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dropNulls}
\alias{dropNulls}
\title{Remove NULL elements from list}
\usage{
dropNulls(x)
}
\arguments{
\item{x}{List to cleanup}
}
\value{
A list.
}
\description{
Remove NULL elements from list
}
\keyword{internal}
