% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_from_list}
\alias{extract_from_list}
\title{Extract selected key from lists}
\usage{
extract_from_list(l, key = "name", type = character(1))
}
\arguments{
\item{l}{List to search in.}

\item{key}{List key to extract. Default to name.}

\item{type}{Needed by \code{vapply} as FUN.VALUE parameter.
Default to character.}
}
\value{
A list.
}
\description{
Extract selected key from lists
}
\keyword{internal}
