% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_evol}
\alias{find_evol}
\title{Recursively find pokemon evolutions}
\usage{
find_evol(tmp, id, entry_point, trigger = NULL, level = NULL, object = NULL)
}
\arguments{
\item{tmp, }{id, entry_point, trigger, level, object Internal parameters}
}
\value{
A list.
}
\description{
When length(tmp$evolves_to) is 0, the function returns
a list with aggregated pokemon names, ids, trigger, levels ...
}
\keyword{internal}
