% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_evols}
\alias{find_evols}
\title{Given a pokemon entry_point return the entire list}
\usage{
find_evols(l)
}
\arguments{
\item{l}{Pokemon evolution chain.}
}
\value{
A list.
}
\description{
Given a pokemon entry_point return the entire list
}
\keyword{internal}
