% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeFightModule.R
\name{generate_pokemons}
\alias{generate_pokemons}
\title{Function that generate 2 random pokemons.
Levels are between 1 and 100 and ids are between
1 and 151}
\usage{
generate_pokemons(mainData, sprites, difficulty, attacks)
}
\arguments{
\item{mainData}{Object containing the main pokemon data.}

\item{sprites}{Object containing pokemon images.}

\item{difficulty}{Game difficulty.}

\item{attacks}{Object containing pokemon attacks.}
}
\description{
Function that generate 2 random pokemons.
Levels are between 1 and 100 and ids are between
1 and 151
}
