% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_front_sprites}
\alias{get_front_sprites}
\title{Get pokemon front sprite}
\usage{
get_front_sprites(shiny = FALSE)
}
\arguments{
\item{type}{Shiny or not Shiny. Boolean, default FALSE.}
}
\description{
Get pokemon front sprite
}
\keyword{internal}
