% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_poke_locations.R
\name{poke_locations_ui}
\alias{poke_locations_ui}
\alias{poke_locations_server}
\title{UI module for generating the pokemon location card}
\usage{
poke_locations_ui(id)

poke_locations_server(id, selected)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{selected}{Input containing the selected pokemon index.}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
UI module for generating the pokemon location card

Server module generating the pokemon location card
}
