% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_poke_moves.R
\name{poke_moves_ui}
\alias{poke_moves_ui}
\alias{poke_moves_server}
\title{UI module for sorting pokemon}
\usage{
poke_moves_ui(id)

poke_moves_server(id, selected)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{selected}{Input containing the selected pokemon index.}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
UI module for sorting pokemon

Server module generating the pokemon interface
}
