% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poke_attacks}
\alias{poke_attacks}
\title{Pokemon attacks data}
\format{
## `who`
A list with `r length(poke_attacks)` elements.
}
\source{
<https://pokeapi.co/docs/v2>
}
\usage{
poke_attacks
}
\description{
Aggregated list of pokemon attacks data
}
\keyword{datasets}
