% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poke_data}
\alias{poke_data}
\title{Pokemon global data}
\format{
## `who`
A list with `r length(poke_data)` elements:
\describe{
  \item{name}{Pokemon name}
  \item{description}{Pokemon description}
  \item{shape}{Pokemon shape}
  ...
}
}
\source{
<https://pokeapi.co/docs/v2>
}
\usage{
poke_data
}
\description{
Aggregated list of pokemon data
}
\keyword{datasets}
