% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_poke_evol.R
\name{poke_evol_ui}
\alias{poke_evol_ui}
\alias{poke_evol_server}
\title{UI module for generating the pokemon evolution timeline}
\usage{
poke_evol_ui(id)

poke_evol_server(id, selected, shiny)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}

\item{selected}{Input containing the selected pokemon index.}

\item{shiny}{Whether to display a shiny version. FALSE by default.}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
UI module for generating the pokemon evolution timeline

Server module generating the pokemon evolution timeline
}
