% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_poke_infos.R
\name{poke_infos_server}
\alias{poke_infos_server}
\title{Server module generating the pokemon info cards}
\usage{
poke_infos_server(id, selected, shiny)
}
\arguments{
\item{id}{Module id.}

\item{selected}{Input containing the selected pokemon index.}

\item{shiny}{Whether to display a shiny version. FALSE by default.}
}
\description{
Server module generating the pokemon info cards
}
