% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poke_network}
\alias{poke_network}
\title{Pokemon network data}
\format{
## `who`
A list with nodes and edges dataframes. For the node dataframe:
\describe{
  \item{id}{Pokemon id}
  \item{label}{Pokemon name}
  \item{group}{Pokemon family}
  \item{image}{Pokemon sprite}
  \item{shape}{Node shape}
  \item{value}{Node value}
}
For edges:
\describe{
  \item{from}{Pokemon previous stage}
  \item{to}{Pokemon next stage}
  \item{title}{Evolution detail}
  \item{arrow}{Direction of arrows}
}
}
\source{
<https://pokeapi.co/docs/v2#evolution-section>
}
\usage{
poke_network
}
\description{
Useful data to reconstruct with visNetwork htmlWidget.
}
\keyword{datasets}
