% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_poke_select.R
\name{poke_select_server}
\alias{poke_select_server}
\title{Server module generating the pokemon interface}
\usage{
poke_select_server(id, selected)
}
\arguments{
\item{id}{Module id.}

\item{selected}{Object containing the selected pokemon in the network, if not NULL.}
}
\description{
Server module generating the pokemon interface
}
