% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_poke_select.R
\name{poke_select_ui}
\alias{poke_select_ui}
\title{UI module for sorting pokemon}
\usage{
poke_select_ui(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
UI module for sorting pokemon
}
