% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_poke_stats.R
\name{poke_stats_server}
\alias{poke_stats_server}
\title{Server module generating the pokemon stats chart}
\usage{
poke_stats_server(id, selected)
}
\arguments{
\item{id}{Module id.}

\item{selected}{Input containing the selected pokemon index.}
}
\description{
Server module generating the pokemon stats chart
}
