% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_poke_stats.R
\name{poke_stats_ui}
\alias{poke_stats_ui}
\title{UI module for generating the pokemon stats chart}
\usage{
poke_stats_ui(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
UI module for generating the pokemon stats chart
}
