% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeAttackModule.R
\name{pokeAttack}
\alias{pokeAttack}
\title{Server module for generating the pokeAttacks section}
\usage{
pokeAttack(input, output, session, attacks)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}

\item{attacks}{Data containing all pokemon abilities.}
}
\description{
Server module for generating the pokeAttacks section
}
