% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeEvolveModule.R
\name{pokeEvolve}
\alias{pokeEvolve}
\title{Server module generating the pokemon evolution timeline}
\usage{
pokeEvolve(input, output, session, mainData, details, selected, shiny)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}

\item{mainData}{Object containing the main pokemon data.}

\item{details}{Object containing extra pokemon details.}

\item{selected}{Input containing the selected pokemon index.}

\item{shiny}{Whether to display a shiny version. FALSE by default.}
}
\description{
Server module generating the pokemon evolution timeline
}
