% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeFight.R
\name{pokeFight}
\alias{pokeFight}
\title{Server module for generating the pokeFight section}
\usage{
pokeFight(input, output, session, mainData, sprites, attacks)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}

\item{mainData}{Object containing the main pokemon data.}

\item{sprites}{Object containing pokemon images.}

\item{attacks}{Object containing pokemon attacks.}
}
\description{
Server module for generating the pokeFight section
}
