% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeFightModule.R
\name{pokeFightUi}
\alias{pokeFightUi}
\title{UI module for generating the pokeFight section}
\usage{
pokeFightUi(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
UI module for generating the pokeFight section
}
