% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeGalleryModule.R
\name{pokeGallery}
\alias{pokeGallery}
\title{Server module generating the pokemon gallery interface}
\usage{
pokeGallery(input, output, session, mainData, details, shiny)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}

\item{mainData}{Object containing the main pokemon data.}

\item{details}{Object containing extra pokemon details.}

\item{shiny}{Whether to display a shiny version. FALSE by default.}
}
\description{
Server module generating the pokemon gallery interface
}
