% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeGalleryModule.R
\name{pokeGalleryUi}
\alias{pokeGalleryUi}
\title{UI module for generating the gallery of the first 151 pokemons}
\usage{
pokeGalleryUi(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
UI module for generating the gallery of the first 151 pokemons
}
