% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeInfosModule.R
\name{pokeInfos}
\alias{pokeInfos}
\title{Server module generating the pokemon info cards}
\usage{
pokeInfos(input, output, session, mainData, details, pokeNames)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}

\item{mainData}{Object containing the main pokemon data.}

\item{details}{Object containing extra pokemon details.}

\item{pokeNames}{Object containing pokemon names.}
}
\description{
Server module generating the pokemon info cards
}
