% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeInfosModule.R
\name{pokeInfosUi}
\alias{pokeInfosUi}
\title{UI module for generating the pokemon info cards}
\usage{
pokeInfosUi(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
UI module for generating the pokemon info cards
}
