% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeInputModule.R
\name{pokeInput}
\alias{pokeInput}
\title{Server module generating the pokemon interface}
\usage{
pokeInput(input, output, session, mainData, sprites, details, selected)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}

\item{mainData}{Object containing the main pokemon data.}

\item{sprites}{Object containing pokemon images.}

\item{details}{Object containing extra pokemon details.}

\item{selected}{Object containing the selected pokemon in the network, if not NULL.}
}
\description{
Server module generating the pokemon interface
}
