% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeDataModule.R
\name{pokeInputUi}
\alias{pokeInputUi}
\title{UI module for sorting pokemon}
\usage{
pokeInputUi(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
UI module for sorting pokemon
}
