% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeLocationModule.R
\name{pokeLocation}
\alias{pokeLocation}
\title{Server module generating the pokemon location card}
\usage{
pokeLocation(input, output, session, mainData, details, selected)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}

\item{mainData}{Object containing the main pokemon data.}

\item{details}{Object containing extra pokemon details.}

\item{selected}{Input containing the selected pokemon index.}
}
\description{
Server module generating the pokemon location card
}
