% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeNetworkInputModule.R
\name{pokeNetworkInputUi}
\alias{pokeNetworkInputUi}
\title{UI module for controling pokeNetwork options}
\usage{
pokeNetworkInputUi(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
UI module for controling pokeNetwork options
}
