% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeOthersModule.R
\name{pokeOther}
\alias{pokeOther}
\title{Server module generating other stats chart}
\usage{
pokeOther(input, output, session, mainData, details)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}

\item{mainData}{Object containing the main pokemon data.}

\item{details}{Object containing extra pokemon details.}
}
\description{
Server module generating other stats chart
}
