% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeOthersModule.R
\name{pokeOtherUi}
\alias{pokeOtherUi}
\title{UI module for generating other stats chart}
\usage{
pokeOtherUi(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
UI module for generating other stats chart
}
