% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeStatsModule.R
\name{pokeStatsUi}
\alias{pokeStatsUi}
\title{UI module for generating the pokemon stats chart}
\usage{
pokeStatsUi(id)
}
\arguments{
\item{id, }{character used to specify namespace, see \code{shiny::\link[shiny]{NS}}}
}
\value{
a \code{shiny::\link[shiny]{tagList}} containing UI elements
}
\description{
UI module for generating the pokemon stats chart
}
