% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeTypesModule.R
\name{pokeType}
\alias{pokeType}
\title{Server module generating the pokemon types info boxes}
\usage{
pokeType(input, output, session, types, selected)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}

\item{types}{Object containing the preprocessed pokemon types.}

\item{selected}{Input containing the selected pokemon index.}
}
\description{
Server module generating the pokemon types info boxes
}
