% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeFightModule.R
\name{select_attacks}
\alias{select_attacks}
\title{Function that list all possible learnable moves for a pokemon
and select only 4 moves.}
\usage{
select_attacks(mainData, attacks, current_pokemon_id)
}
\arguments{
\item{mainData}{Object containing the main pokemon data.}

\item{attacks}{Object containing all pokemon attacks data.}

\item{current_pokemon_id}{Id of the randomly generated pokemon.}
}
\description{
Function that list all possible learnable moves for a pokemon
and select only 4 moves.
}
\note{
Limitations: only physical moves are considered for the moment. This function
has to be called inside \link{generate_pokemons}.
}
