% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeFightModule.R
\name{fight_History}
\alias{fight_History}
\title{Function that generates a timelineItem for each attack the pokemon does.}
\usage{
fight_History(attacking, opponent, attacks, current_attack, types)
}
\arguments{
\item{attacking}{Id of the randomly generated pokemon who is attacking.}

\item{opponent}{Opponent type. Useful for effectiveness calculations.}

\item{attacks}{Object containing all pokemon attacks data.}

\item{current_attack}{The currently selected attack.}

\item{types}{Object containing all pokemons types strenght and weaknesses.
TRUE in this case.}
}
\description{
Function that generates a timelineItem for each attack the pokemon does.
}
