% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeGalleryModule.R
\name{pokeGallery}
\alias{pokeGallery}
\title{Server module generating the pokemon interface}
\usage{
pokeGallery(input, output, session, raw_data, raw_details)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}

\item{raw_data}{Object containing the main pokemon data.}

\item{raw_details}{Object containing extra pokemon details.}
}
\description{
Server module generating the pokemon interface
}
