% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeDataModule.R
\name{pokeInput}
\alias{pokeInput}
\title{Server module generating the pokemon interface}
\usage{
pokeInput(input, output, session, mainData, details)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}

\item{mainData}{Object containing the main pokemon data.}

\item{details}{Object containing extra pokemon details.}
}
\description{
Server module generating the pokemon interface
}
