% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeLocationModule.R
\name{pokeLocation}
\alias{pokeLocation}
\title{Server module generating the pokemon location card}
\usage{
pokeLocation(input, output, session, selected, locations)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}

\item{selected}{Input containing the selected pokemon index.}

\item{locations}{Contains preprocessed data of the selected pokemon location}
}
\description{
Server module generating the pokemon location card
}
