% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeMovesModule.R
\name{pokeMove}
\alias{pokeMove}
\title{Server module generating the pokemon interface}
\usage{
pokeMove(input, output, session, selected, moves)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}

\item{selected}{Input containing the selected pokemon index.}

\item{moves}{Contains preprocessed pokemon moves.}
}
\description{
Server module generating the pokemon interface
}
