% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeNetworkModule.R
\name{pokeNetwork}
\alias{pokeNetwork}
\title{Server module for generating the pokeNetwork section}
\usage{
pokeNetwork(input, output, session, mainData)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}

\item{attacks}{Data containing all pokemon abilities.}
}
\description{
Server module for generating the pokeNetwork section
}
