% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeNetworkModule.R
\name{pokeNetwork}
\alias{pokeNetwork}
\title{Server module for generating the pokeNetwork section}
\usage{
pokeNetwork(
  input,
  output,
  session,
  mainData,
  details,
  families,
  groups,
  mobile,
  networkOptions
)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}

\item{mainData}{All pokemon main data.}

\item{details}{Object containing extra pokemon details.}

\item{families}{List containg all pokemon connections.}

\item{groups}{List containing data for grouping pokemons by evolution family.}

\item{mobile}{Shiny input checking if the app is running on a cellphone/tablet.}

\item{networkOptions}{Slot for the pokeNetworkInputModule inputs.}
}
\description{
Server module for generating the pokeNetwork section
}
