% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeNetworkModule.R
\name{pokeNetwork}
\alias{pokeNetwork}
\title{Server module for generating the pokeNetwork section}
\usage{
pokeNetwork(input, output, session, mainData, families, groups, mobile)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}

\item{mainData}{All pokemon main data.}

\item{families}{List containg all pokemon connections.}

\item{groups}{List containing data for grouping pokemons by evolution family.}

\item{mobile}{Shiny input checking if the app is running on a cellphone/tablet.}
}
\description{
Server module for generating the pokeNetwork section
}
