% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeNetworkInputModule.R
\name{pokeNetworkInput}
\alias{pokeNetworkInput}
\title{Server module for controling the pokeNetwork section}
\usage{
pokeNetworkInput(input, output, session)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}
}
\description{
Server module for controling the pokeNetwork section
}
