% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeStatsModule.R
\name{pokeStats}
\alias{pokeStats}
\title{Server module generating the pokemon stats chart}
\usage{
pokeStats(input, output, session, skills, pokeNames, selected)
}
\arguments{
\item{input}{Shiny inputs.}

\item{output}{Shiny outputs.}

\item{session}{Shiny session.}

\item{skills}{Object containing pokemon statistics.}

\item{pokeNames}{Object containing pokemon names.}

\item{selected}{Input containing the selected pokemon index.}
}
\description{
Server module generating the pokemon stats chart
}
